# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "api para ver el tema del coso delcoso sobre el cosito(es de comprar cosas)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "LA api"
        }
    ],
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "todo lo pokemone",
                "operationId": "6e8baa4c91af0d3c2642318b1d6167ce",
                "responses": {
                    "200": {
                        "description": "aqui etan"
                    },
                    "401": {
                        "description": "no eta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "uno mas",
                "operationId": "ad14be70e08e012e93c318b31aef94c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut",
                                    "nombre",
                                    "apellido",
                                    "correo",
                                    "contraseña"
                                ],
                                "properties": {
                                    "rut": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@ventasfix.cl"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "example": "Clave123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "spawneo el compa"
                    },
                    "422": {
                        "description": "spawneo mal parece"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/usuarios/{usuario}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "dime tu edad y sabre que edad tienes",
                "operationId": "5d9adbdba1510bf02c6b6e4e2ecc3186",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aqui eta el chavo"
                    },
                    "404": {
                        "description": "se perdio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "tamo actualizando gratis",
                "operationId": "c8c1c3ac9081879bd9fcbe30573e3cbc",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut",
                                    "nombre",
                                    "apellido",
                                    "correo"
                                ],
                                "properties": {
                                    "rut": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@ventasfix.cl"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "example": "(opcional) Clave123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ya tamo, vuelva nunca"
                    },
                    "404": {
                        "description": "y ete quien e"
                    },
                    "422": {
                        "description": "a nose que paso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "aqui se expulsa del clan a la gente",
                "operationId": "abf6ef762a79aba030e0275b3053691e",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "chau pai"
                    },
                    "404": {
                        "description": "y ete quien e"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "el token de segiruda",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}